run tests in multi module
