+++ another_module/.gitignore
