+++ another_module/proguard-rules.pro
