+++ another_module/src/androidTest/java/br/com/another/AnotherModuleActivityTest.java
@@ -1,0 +1,27 @@
+package br.com.another;
+
+import android.support.test.rule.ActivityTestRule;
+import android.support.test.runner.AndroidJUnit4;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.click;
+import static android.support.test.espresso.assertion.ViewAssertions.matches;
+import static android.support.test.espresso.matcher.ViewMatchers.*;
+
+@RunWith(AndroidJUnit4.class)
+public class AnotherModuleActivityTest {
+
+    @Rule
+    public ActivityTestRule<AnotherModuleActivity> rule = new ActivityTestRule<>(AnotherModuleActivity.class);
+
+    @Test
+    public void shouldUpdateTextAfterButtonClick() {
+        onView(withId(R.id.button)).perform(click());
+
+        onView(withId(R.id.button)).check(matches(withText("hello")));
+    }
+}
