+++ another_module/src/main/AndroidManifest.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="br.com.another">
+
+    <application
+        android:theme="@style/AppTheme">
+        <activity android:name=".AnotherModuleActivity"></activity>
+    </application>
+
+</manifest>
