+++ another_module/src/main/java/br/com/another/AnotherModuleActivity.java
@@ -1,0 +1,19 @@
+package br.com.another;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+public class AnotherModuleActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_another_module);
+    }
+
+    public void onClick(View view) {
+        ((Button)view).setText("hello");
+    }
+}
