+++ another_module/src/main/kotlin/br/com/another/AnotherKotlinClass.kt
@@ -1,0 +1,6 @@
+package br.com.another
+
+class AnotherKotlinClass {
+
+    fun multiply(first: Int, second: Int) = first * second
+}
