+++ another_module/src/main/res/values/strings.xml
