+++ app/src/main/res/layout/activity_main.xml
@@ -22,4 +22,11 @@
         android:layout_height="wrap_content"
         android:text="Don't Click Me!" />
 
+    <Button
+        android:id="@+id/navigate"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Navigate" />
+
+
 </LinearLayout>
