--- kotlin/src/main/kotlin/br/com/tramalho/AnotherKotlinClass.kt
@@ -1,6 +1,0 @@
-package br.com.tramalho
-
-class AnotherKotlinClass {
-
-    fun multiply(first: Int, second: Int) = first * second
-}
