--- kotlin/src/test/kotlin/br/com/tramalho/MultiplyTest.kt
@@ -1,14 +1,0 @@
-package br.com.tramalho
-
-import org.hamcrest.CoreMatchers.`is`
-import org.hamcrest.MatcherAssert.assertThat
-import org.junit.Test
-
-class MultiplyTest {
-
-    @Test
-    fun checkIfMultiplyIsCorrect() {
-
-        assertThat(AnotherKotlinClass().multiply(2, 1), `is`(2))
-    }
-}
